requirements.txt
whisper-openai
tqdm